1.6.16a
- date column - include in sorting
- goto -> show visually where going to (like, after 10 lines, before x
lines, to date XXX, extra x seconds etc)
- goto date -> works as long as it's sorted ascending (in event viewer,
by default it's decending)

+++ src/lw_common/ui/log_view/log_view.cs
@@ -1409,7 +1409,7 @@ namespace lw_common.ui
                 go_to_row(closest.Item2, select_type.notify_parent);
         }
 
-        public void offset_closest_time(int time_ms, bool forward) {
+        public void offset_closest_time(long time_ms, bool forward) {
             if (filter_.match_count < 1)
                 return;
             int sel = sel_row_idx;
